server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai-agents[voice]>=0.3.2",
    "ruff>=0.13.2",
    "python-dotenv>=1.0.1",
    "dotenv>=0.9.9",
    "uvicorn>=0.37.0",
    "fastapi>=0.117.1",
    "websockets>=15.0.1",
]
